+++ pedroguerra/src/main/java/com/pedroguerra/model/RelatorioServico.java
@@ -1,16 +1,17 @@
 package com.pedroguerra.model;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
 public class RelatorioServico {
     private String fkServicoId;
-    private float area;
+    private BigDecimal area;
     private Date dataRelatorio;
     private String observacoes;
 
     public RelatorioServico() {}
 
-    public RelatorioServico(String fkServicoId, float area, Date dataRelatorio, String observacoes) {
+    public RelatorioServico(String fkServicoId, BigDecimal area, Date dataRelatorio, String observacoes) {
         this.fkServicoId = fkServicoId;
         this.area = area;
         this.dataRelatorio = dataRelatorio;
@@ -20,8 +21,8 @@ public class RelatorioServico {
     public String getFkServicoId() { return fkServicoId; }
     public void setFkServicoId(String fkServicoId) { this.fkServicoId = fkServicoId; }
 
-    public float getArea() { return area; }
-    public void setArea(float area) { this.area = area; }
+    public BigDecimal getArea() { return area; }
+    public void setArea(BigDecimal area) { this.area = area; }
 
     public Date getDataRelatorio() { return dataRelatorio; }
     public void setDataRelatorio(Date dataRelatorio) { this.dataRelatorio = dataRelatorio; }
